chore(files): move tests to tests folder
