+++ src/tests/App.test.tsx
@@ -1,5 +1,5 @@
 import { render, screen } from '@testing-library/react';
-import App from './App';
+import App from '../App';
 
 test('renderizar componente Button com texto', () => {
   render(<App />);
